Added separate controller for saving document test results
