+++ src/server/java/uk/gov/dvsa/mot/server/controller/DataController.java
@@ -1,19 +1,15 @@
 package uk.gov.dvsa.mot.server.controller;
 
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
-import static org.springframework.web.bind.annotation.RequestMethod.POST;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.dvsa.mot.server.data.DatabaseDataProvider;
-import uk.gov.dvsa.mot.server.reporting.DocumentListReportGenerator;
 
 import java.util.List;
 import javax.inject.Inject;
@@ -60,41 +56,5 @@ public class DataController {
     public ResponseEntity<List<String>> getCachedDatasetEntry(@PathVariable String datasetName) {
         logger.info("getCachedDatasetEntry: name {}", datasetName);
         return ResponseEntity.ok(dataProvider.getCachedDatasetEntry(datasetName));
-    }
-
-    /**
-     * Get the timestamp used by DocumentListReportGenerator.
-     * @return timestamp
-     */
-    @RequestMapping(value = "/timestamp", method = GET,
-            produces = MediaType.TEXT_PLAIN_VALUE)
-    public ResponseEntity<String> getTimestamp() {
-        logger.info("getTimestamp");
-        return ResponseEntity.ok(DocumentListReportGenerator.getTimestamp());
-    }
-
-    /**
-     * Add a document result.
-     * @param result to add.
-     * @return confirmation
-     */
-    @ResponseBody
-    @RequestMapping(value = "/documents/results", method = POST,
-            produces = MediaType.TEXT_PLAIN_VALUE)
-    public ResponseEntity<String> addDocumentResults(@RequestBody String result) {
-        DocumentListReportGenerator.addDocumentResult(result);
-
-        return ResponseEntity.ok(String.format("Added document result: %s.", result));
-    }
-
-    /**
-     * Get all document results.
-     * @return document results.
-     */
-    @ResponseBody
-    @RequestMapping(value = "/documents/results", method = GET,
-            produces = MediaType.TEXT_PLAIN_VALUE)
-    public ResponseEntity<String> getDocumentResults() {
-        return ResponseEntity.ok(DocumentListReportGenerator.getDocumentResults());
     }
 }
